[xinit] add xbindkeys support
Former-commit-id: b1e1e07e3be06a0c9a63f18d5fd991c73180a0b9 [formerly c2264e5a3cd9597bb740e7d89abe64186f5393d6] [formerly 2a7bc25ba2c5b9279f95418dbf0a94c0044d1e43 [formerly f47c22b872ead854ffc319eb4c26153c3d16c925]]
Former-commit-id: d003808864a998c40a0253ffe11579fc44d4bf28 [formerly f2af83c47fbcc9de7ca6415c6c4cabb5853a2633]
Former-commit-id: 2ed778f9f5284b72246a9f88510d2b3bb5f5c139

+++ .config/xinit/hotkeys.zsh
@@ -1,2 +1,7 @@
-rm -f "${SXHKD_FIFO}" && mkfifo "${SXHKD_FIFO}" && \
-    nexec sxhkd && sxhkd -s "${SXHKD_FIFO}" > /dev/null &
+USE_SXHKD=${1:-true}
+if [[ $USE_SXHKD ]]; then
+    rm -f "${SXHKD_FIFO}" && mkfifo "${SXHKD_FIFO}" && \
+        nexec sxhkd && sxhkd -s "${SXHKD_FIFO}" > /dev/null &
+else
+    xbindkeys -p -n > /tmp/xbindkeys$$.log
+fi
